[back-end] src/service.js - implement routes
Give them a stub implementation which is backed by files for now. So
the API endpoint parameters are mapped to filenames, and these are
found in a directory (be default 'data/' relative to the PWD, but this
can be overridden by an environment variable SERVER_DATA_ROOT) and
regurgitated verbatim.