deploy.sh - a script for deploying the application
Should be run as the user running the service. Manages the server
process with a systemd service installed in user-mode to ensure that
it restarts on boot.

Requires proxying and web directories to configured beforehand, as
this requires elevated privileges.

ESM/CSM build glitches avoided by not building source more than
minimally with vite, and running direct from the project.

This may need attention for cleaning up later.

Runs backed via this run-script:

    npm run start:attached