[front-end] server.ts etc. - nominally working unified vite server
"Unified" in the sense that the vite server runs the back-end API as
well as serving the front end, which allows the front end to be
launched in development with a working back-end. The back end doesn't
need to be launched on a separate HTTP server and CORS configured to
allow it to connect (although that can still be done by configuring
FASTIFY_CORS_ORIGIN=http://localhost:5173 in the back-end's .env
file).

Vite still rebuilds the front end dynamically, although for reasons
I've not tracked down, it doesn't open a new browser window on
launch. Deferring addressing that to a future commit.

Also deferring adding an automated rebuild of the back-end to a future
commit.

Note, we now need two tsconfig files, one for the front end browser
code, one for the nodejs code that runs in server.ts

The TS code gets transpiled into the node-dist/ directory, which we
.gitignore (/dist is already ignored)

We alter the run script target `dev` for building and running this
unified server. This now does what `dev` did, and more (it runs the
backend; however it doesn't open a browser window, you need to do that
yourself, but this is a relatively small thing). As it embeds the vite
rebuild process, hot module reloading should still work, whilst also
working in production without vite.

Note that the server.ts requires process.env.NODE_ENV to be
"development" for it to work, this seems to be a bug in vite or one of
the related vite tools, but essentially we won't normally be running
this target in production mode. Leaving this dev option settable by
NODE_ENV for now, in case we can fix this later to allow running this
with NODE_ENV=production, which could be usefull for checking some
things.

Note, if we do need what used to be done by the `dev` runscript, the
`start` runscript still exists and is identical to it.

Note we set the vite root option from the current working directly -
this assumes npm has been run from project root, but this is usually
valid (if not will need to adapt this later).

The reason: deducing the root path from the config file location has
been breaking when the location is changing due to TS compilation to
JS