[back-end] src/server.ts - don't start server by default we we can be middleware
We will be consumed as a library later, when the front-end development
server runs this server as middleware. Therefore we don't want to
start the server by default here. Instead, export the start function,
so that importers can perform the start themselves. Also export the
API plugin, and the options associated with that.