[back-end] typescriptification - remove // @ts-nocheck from source
And fill in the missing types.

Fastify's type definitions seem super complicated - sidestepping
engaging too deeply with this.  Our module seems fairly simple,
however.

Some details of this here:

https://fastify.dev/docs/latest/Reference/TypeScript/#creating-a-typescript-fastify-plugin

NOTE: following that guide, in pluginApi.ts I am now returning
`fp(pluginApi)` rather than just returning `pluginApi`. I'm not clear
on why this didn't need to be done before or if/why it's needed now,
but a consequence of this is that the "doubling up" problem where the
`prefix` option needs to be deleted to avoid it being processed twice,
is now gone. So that code needs to be deleted, otherwise the prefix
option doesn't get handled at all!

fixup post typescriptify-back-end