[back-end] vite.config.ts, etc. - add __BUILD_INFO__ define
This gets inserted statically by the vite build into the project
source. It requires some typing declaration help in src/app.d.ts, and
some newer ES2022 definitions in tsconfig.json.

The __BUILD_INFO__ contains information gleaned at build time about
the state of the build - notably what version was built, when, and
with what node environment.

The intention is to expose this via the API in a later commit.