[common] contract.ts - add validations for ConfigData type
Ready for getConfig endpoint, which will be added in an upcoming
commit (it will affect the front-end and back end)

Note, we can remove the anchors in the QName regex as ZodRegex adds
those for us. This makes all the regex definitions consistently
non-anchored, which makes them composable into other regexes.