[common] test/file-utils.ts - currently just defines slurp() variations
"slurp" is a function name (and an idea) I stole from Perl's
ecosystem.

These functions are intended to make loading a file a one-call
operation, and so to make loading files in tests shorter. I'd have
included some globbing functionality here except there is a problem
inferring the caller's source directory from inside a function, which
makes it seem less useful (so that feature is now going into a "glorp"
function in the tests themselves, committed later). The functions may
also be useful more generally.

I'm putting these in the test/ directory as they're NodeJS specific,
so don't really belong in a common module shared with the front-end.
If the back-end tests (or even the back-end implementation) wants to
use them, we will need to move them to another common library which
can target NodeJS.