[back-end] bin/dataset.ts, package.json - the importer script
Run it with

   npm run dataset <parameters here>

Currently we have just one subcommand `import`: reads in a CSV file
and write the corresponding back-end data to a directory.

The implementation is in `src/dataset.ts` - `bin/dataset.ts` is just
an executable. This is to allow us to unit test the implementation
without having to import the executable, which would run the
implementation immediately in an uncontrolled way.

The CSV data rows are converted to DatasetItem objects. The
DatasetItem.id property is mandatory, and duplicates are detected.

A `config.json` file needs to be supplied to the script, and this
should define the properties to put in the data. Also, it defines the
CSV headers each property should be imported from using the (otherwise
optional) `from` field. The implementation is fairly simple, and no
validation is performed yet, apart from requiring a unique ID
field. The vocabs can also be supplied in this file; this may be used
to validate data later.