[common] package.json - add dist/ to `npm run clean`
Without this the compiled contract doesn't get removed which can play
havoc with everything else (backend tests crashing as below etc.)

```

⎯⎯⎯⎯⎯ Uncaught Exception ⎯⎯⎯⎯⎯
TypeError: Cannot read properties of undefined (reading 'method')
 ❯ registerRoute ../../node_modules/@ts-rest/fastify/index.esm.mjs:140:26
 ❯ recursivelyRegisterRouter ../../node_modules/@ts-rest/fastify/index.esm.mjs:214:9
 ❯ recursivelyRegisterRouter ../../node_modules/@ts-rest/fastify/index.esm.mjs:209:13
 ❯ ../../node_modules/@ts-rest/fastify/index.esm.mjs:97:9
 ❯ Plugin.exec ../../node_modules/avvio/lib/plugin.js:125:28
 ❯ Boot._loadPlugin ../../node_modules/avvio/boot.js:432:10
 ❯ processTicksAndRejections node:internal/process/task_queues:82:21

This error originated in "test/plugin.test.ts" test file. It doesn't mean the error was thrown inside the file itself, but while it was running.
The latest test that might've caused the error is "testing datasetItem". It might mean one of the following:
- The error was thrown, while Vitest was running this test.
- If the error occurred after the test had been completed, this was the last documented test before it was thrown.
```