[common] src/utils.ts - some useful functions
This imported wholesale, if somewhat tidied, from the previous Mykomap
project. I'm not going to weed them out right now - it's possible they
may not turn out to be useful but they aren't hugely heavy to carry.

The rationale for this is that having these also makes importing other
pieces of code from Mykomap-1 a bit easier, because they often rely on
these definitions.