[back-end] src/csv.ts - utilities for streaming CSV files
We use udsv, which looks like the fastest CSV parser in town right
now.  MM 3.x used to use PapaParse, but that module is tailored
specifically for use in the browser, and is hard to run on the nodejs
platform. It's also benchmarked as slower, according to udsv's project
page.

In this implementation, CSV rows are translated into some arbitrary
format (or alternatively, perhaps left as a string array) by a
transformer function. This function is generated when the CSV headers
become available, by a callback the caller supplies, Any validation of
the headers should be done in this callback.

In practice, the CSV rows will be translated into DatasetItems. This
is done in an upcoming commit.