[common] src/prop-specs.ts - property specifiers
These are interfaces, and allow short literal objects to be used to
specify property definitions.  They're useful for JSON or config
literals.

Adopted mostly wholesale from Mykomap 3.x, except that to avoid the
complexity of using pure interfaces experienced in MM 3.x, they have
been renamed as PropSpecs so that the name PropDefs can be used for
class implementations, that can package the logic in a form that
avoids switch conditions. (In an upcoming commit).

Another change is to use FilterSpec to denote filter presets rather
than a non-undefined value, and absence indicating no filter. The
latter (absent vs undefined properties) doesn't seem to be visible to
typescript typing for classes, so it emerges we can't use it in the
way we did if we want class implementations of the logic.