[common] src/prop-defs.ts - property definitions
These are the classes which add the logic which was present in Mykomap
3.x but was somewhat awkward to implement and use when implemented
with pure interfaces - lots of conditionals and switch statements were
required. With classes, the choice of logic is implied by the class
heirarchy.

Instances of these are what you want to be using when operating on
property definitions. Whereas, PropSpecs are still useful for JSON and
literals.

The classes use the equivalent PropSpecs as contructor
parameters (which could be other PropDefs, or JSON).