[front-end] minimal changes to allow compilation with new ConfigData contract
We need to change `fields` to be `itemProps` for the definitions to be
consistent without duplications. Short-hand `"value"` field
definitions need to be expanded, as that isn't supported right (this
seemed a low priority).

Some usages of the word "field" in the code has been changed to "prop"
for consistency.

I've not removed the `filterableFields` in the `Config` type. The new
`itemProps` element of ConfigData has the filter options which were
added in recent MM 3.x versions included (when `filterableFields` was
deprecated), and these probably should be used instead. But this is
outside the scope of this PR.

Aside: I recommend constucting PropDef class instances rather than
using PropSpec interfaces in general, as this makes introspection much
easier. The class constructors can be supplied with PropSpecs as
parameters, but they will need vocab definitions. PropertyServices is
a factory class which aims to make it easier to construct them in
bulk, given common vocab definitions.

Note, this is all essentially slightly adapted code from Mykomap v3.x,
and not a completely new invention. In MM 3.x it was in a class called
DataServices, which included a lot more besides, which hasn't been
included here.