[common] src/api/contract.ts - add AbbrevUri validator for VocabPropSpec's URI
The VocabPropSpec.uri field needs to be an abbrevUri, i.e. have a
trailing colon with no second NCname, rather than a full QName,
otherwise we can't reference the vocabs as they are defined in the
ConfigData structure currently.

Probably we shouldn't allow them to be PrefixUris, even though this
could be done, it is just a verbose way of doing the same thing (since
the prefix URI can be inferred from the prefix index)