[common] src/utils.ts - add notNullish and notUndefined functions
These allow removing nulls and undefined values from iterable objects
like arrays, using the .filter() method, without needing to specify
the type in an `is Blah`. Useful when the type is longwinded!

e.g. given:

   const x: Array<Some<Very<Complicated, Combination>, Of<P, Q, R>>|null> = ....

We can do this to get a null-free array typed appropriately:

   x.filter(notNullish)

Instead of needing to do this:

   x.filter(
     (it): is Some<Very<Complicated, Combination>, Of<P, Q, R>> =>
       it != null
   )