[front-end] package.json - tweak tsx invocation for start runscript
The watch was restarting the server when emacs creates lockfiles, and
when vite creates temp configs (the latter causes cyclic
restarts). Not nice.

Note, I'm using the tsx-specific "watch" mode here now instead of the
generic node `--watch` option, as the tsx manual says that's more
robust. It does seem to work better. It also logs what it's doing,
which is helpful.