mapLibre.ts - stop the popups from moving around when panning
MapLibre pop-ups by default try to resist going off-screen when
panned.  This causes a bit of havoc when small-screen users try to
pan, because the pop-up jumps around unpredicably.

But if we specify where to put the anchor, the pop-up will stick with
that and stay still with respect to the tiling.