[back-end] src/index.ts - refactor the start function into a Launcher class
This seems the simplest way to allow Fastify application contruction
to be simple in the default case (when the server is run alone), but
also allow injection of other paramters and plugins (when the server
is run as a FastifyVite plugin on the front-end).

Launcher.start() does what start() did. The start.ts and start.js
scripts need updating to use that.

But you can call the individual methods and tweak things in between if
you need to.