[common] build-info.ts - add the common build definitions here
These will define the build labelling used in this and other modules.

However, they need not to impose the use of Node. Therefore, although
we need to define how to call `git-describe`, we delegate the actual
invocation to an externally supplied function (which can be supplied
at build time and/or where necessary, typically implemented with
`spawnSync`).

We also wrap sentryDist and sentryRelease with as similarly-named
methods for convenience of use.