**/* - npm run prettier:format -> cosmetic changes.
We use prettier.js to enforce formatting and indentation, with the
intention of keeping formatting styles conistent, thereby avoiding
churn when one person's style overrides another's.

However, a lot of non-prettierfied code seems to have built up.  And
ironically, the likely outcome of that is that prettier will obfuscate
future commits when it steps in and reformats entire files in which
only one part is changed - resulting in cosmetic changes getting mixed
into and obscuring the actual functional changes.

Therefore, now seems like a good time to impose the churn of enforcing
to the prettier formatting standards in the project universally, to
avoid them creeping in later.