[common] tsconfig.json - don't include vite.config.ts
This is an attempt to get rid of what seems to be a spurious build
error reporting that:

```
vite v6.3.3 building for production...
vite.config.ts:1:10 - error TS2305: Module '"vite"' has no exported member 'defineConfig'.

1 import { defineConfig } from "vite";
           ~~~~~~~~~~~~
```

This function definitely is defined. vite.config.ts is a bit special
however, in that it is compiled by Vite out of the normal flow -
perhaps this is why?

In any case, this gets rid of the error - not sure if it will cause
problems elsewhere, let's wait and see.