[common] prop-defs.ts - add an uri property to CommonPropDef
This is so that, given an arbitrary prop-def instance, it's less work
to find out if there is an `uri` property and get it if there is.

We often need to do this, when dealing with vocab properties, which
could be simple VocabPropDefs or MultiPropDefs of VocabPropDefs.

To do it:

    // if this is undefined, the propDef isn't a vocab. Otherwise, it
    // is and we get the uri
    const uri = propDef.uri;