[back-end] dataset.ts,cli/dataset.ts - get marker indexes from vocab terms
Adjust DatasetWriter to have a default markerIndex method returning
undefined. Use that to look up the marker index to insert, if
something defined is returned.

Then when generating data for the dataset CLI command, if
config.ui.marker_property_name is set, and it references a vocab
propdef, then override this method to get the index.

It gets this index using the terms from the corresponding vocab
definition for the named property as a mapping to marker index
numbers. The index is the index of the property, or undefined if the
property isn't defined, or -1 if there's a mismatch.  This will get
mapped to zero.

If that property isn't set to anything, the default is as before:
don't add a marker index number.