[back-end] services/Dataset.ts - pagination clarification
The code in question here was a bit obscure and prompted me to go and
test it was actually working correctly. This is a suggested alteration
which I think might be worth including to make the intent clearer to
newcomers, potentially avoiding the investigation I made.

Having run and inspected the unit tests tests, and adding some more
cases, I found it does appear to be performing as designed. But not
obviously so, partly because the design actually intends to do
something surprising (return all the items instead of a 400 error if
page or pageSize are absent). At least, surprising to me, having read
the code and the comments.

The modified code also passes all the tests.