[common] index.ts - export regexes, utilities, inferred contract types
These are useful elsewhere. The inferred types are being re-inferred a
lot - we can just infer them once.

Adjust the previously exported NCName type, which is just an alias for
a string and so not that useful, to be the regex pattern instead. The
main reason for this is to keep the export declarations simple, by
avoiding the conflicting definitions.