[common] rxdefs.ts - define some "uaon" matchers and parseAbbrevUri
The chief aim here is to add a parser for AbbrevUri, which both
validates and returns the NCname term before the colon. And which
makes parsing AbbrevUri values semanically clearer (just a split on
":" doesn't really communicate that, nor distinguish parsing of a
QName from an AbbrevUri, nor catch invalid cases.)

To do that we need an AbbrevUri validator regex.

The regular expressions previously defined here are by design intended
to match substrings. So we need to use the `uaon` (unicode
all-or-nothing) function to convert them to a full-string matcher.

Whilst we're here, let's add some for the other cases we might be
parsing later, in the UAON object.