[back-end] cli/dataset.ts - deal with multi-value properties
There is an implicit assumption that the property named by
ui.marker_property_name is single-valued. If it is not, there will be
an error.

It may not be. So handle these cases, in a rudimentary way:
- If there are no values, treat this as if it were an undefined value.
- If there is one value, treat this as for the single-valued case.
- If there are more than that, return an index one more than the largest

In this fashion, they sort of still work, and the multi-valued case
can be represented with a special icon. (Later, at least, when the
specific icons can be defined.)

An alternative would be to detect the multi-value case and throw an
informative exception rather than just erroring out, but since we sort
of do want to support multi-valued properties for Powys (and sort of
don't as I now think we should adjust the data solve the problem
without them), let's keep this adjustment in now it's written.