[front-end] components/map/markers.ts - fixup marker mapping
Problem: food_system_category is a vocab with 19 terms, some of which
are top-level terms which have markers. But if only 5 mappings are
supplied, this doesn't supply a mapping for all those terms, nor even
a vaguely correct one.

Resolution: We need to supply mappings for all of the terms, not just
the top-level ones. (Or we need to split the vocab into two vocabs,
and deal with the resulting complication, currently hoping to avoid
that)